refactor(ios): split weight page into separate files and add edit/delete
- Extract WeightChartCard, WeightRecordsList, CreateWeightSheet into
  separate files from the monolithic WeightPageView.swift
- Add WeightDetailSheet with view/edit/delete functionality
  (backend already supports POST /weights/update/{id} and /delete/{id})
- WeightPageView is now a slim composition-only file
- WeightStatsRow stays inline in WeightPageView (12 lines)
- Tap a weight record to open detail sheet with edit and delete actions

Closes #125
